backend/requirements.txt
fastapi==0.115.0
uvicorn==0.30.6
httpx==0.27.2
PyJWT==2.9.0
python-dotenv==1.0.1
sse-starlette==2.1.3
pydantic-settings==2.5.2
